requirements.txt
flask
Click==7.0
cmake==3.13.2.post1
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.15.4
Pillow==5.4.1
